rename command to gendocs and make gen-docs
