+++ docs/mctl.md
@@ -1,0 +1,24 @@
+## mctl
+
+mctl is a CLI utility to interact with metal toolbox services
+
+### Options
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+  -h, --help            help for mctl
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl completion](mctl_completion.md)	 - Generate the autocompletion script for the specified shell
+* [mctl create](mctl_create.md)	 - Create resources
+* [mctl delete](mctl_delete.md)	 - Delete resources
+* [mctl edit](mctl_edit.md)	 - Edit resources
+* [mctl gendocs](mctl_gendocs.md)	 - Generate markdown docs for mctl CLI
+* [mctl get](mctl_get.md)	 - Get resource
+* [mctl install](mctl_install.md)	 - Install actions
+* [mctl list](mctl_list.md)	 - List resources
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
