+++ docs/mctl_completion.md
@@ -1,0 +1,32 @@
+## mctl completion
+
+Generate the autocompletion script for the specified shell
+
+### Synopsis
+
+Generate the autocompletion script for mctl for the specified shell.
+See each sub-command's help for details on how to use the generated script.
+
+
+### Options
+
+```
+  -h, --help   help for completion
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl](mctl.md)	 - mctl is a CLI utility to interact with metal toolbox services
+* [mctl completion bash](mctl_completion_bash.md)	 - Generate the autocompletion script for bash
+* [mctl completion fish](mctl_completion_fish.md)	 - Generate the autocompletion script for fish
+* [mctl completion powershell](mctl_completion_powershell.md)	 - Generate the autocompletion script for powershell
+* [mctl completion zsh](mctl_completion_zsh.md)	 - Generate the autocompletion script for zsh
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
