+++ docs/mctl_completion_bash.md
@@ -1,0 +1,51 @@
+## mctl completion bash
+
+Generate the autocompletion script for bash
+
+### Synopsis
+
+Generate the autocompletion script for the bash shell.
+
+This script depends on the 'bash-completion' package.
+If it is not installed already, you can install it via your OS's package manager.
+
+To load completions in your current shell session:
+
+	source <(mctl completion bash)
+
+To load completions for every new session, execute once:
+
+#### Linux:
+
+	mctl completion bash > /etc/bash_completion.d/mctl
+
+#### macOS:
+
+	mctl completion bash > $(brew --prefix)/etc/bash_completion.d/mctl
+
+You will need to start a new shell for this setup to take effect.
+
+
+```
+mctl completion bash
+```
+
+### Options
+
+```
+  -h, --help              help for bash
+      --no-descriptions   disable completion descriptions
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl completion](mctl_completion.md)	 - Generate the autocompletion script for the specified shell
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
