+++ docs/mctl_completion_zsh.md
@@ -1,0 +1,53 @@
+## mctl completion zsh
+
+Generate the autocompletion script for zsh
+
+### Synopsis
+
+Generate the autocompletion script for the zsh shell.
+
+If shell completion is not already enabled in your environment you will need
+to enable it.  You can execute the following once:
+
+	echo "autoload -U compinit; compinit" >> ~/.zshrc
+
+To load completions in your current shell session:
+
+	source <(mctl completion zsh)
+
+To load completions for every new session, execute once:
+
+#### Linux:
+
+	mctl completion zsh > "${fpath[1]}/_mctl"
+
+#### macOS:
+
+	mctl completion zsh > $(brew --prefix)/share/zsh/site-functions/_mctl
+
+You will need to start a new shell for this setup to take effect.
+
+
+```
+mctl completion zsh [flags]
+```
+
+### Options
+
+```
+  -h, --help              help for zsh
+      --no-descriptions   disable completion descriptions
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl completion](mctl_completion.md)	 - Generate the autocompletion script for the specified shell
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
