+++ docs/mctl_create.md
@@ -1,0 +1,28 @@
+## mctl create
+
+Create resources
+
+```
+mctl create [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for create
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl](mctl.md)	 - mctl is a CLI utility to interact with metal toolbox services
+* [mctl create firmware](mctl_create_firmware.md)	 - Create firmware
+* [mctl create firmware-set](mctl_create_firmware-set.md)	 - Create a firmware set
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
