+++ docs/mctl_create_firmware-set.md
@@ -1,0 +1,29 @@
+## mctl create firmware-set
+
+Create a firmware set
+
+```
+mctl create firmware-set [flags]
+```
+
+### Options
+
+```
+      --firmware-uuids string   comma separated list of UUIDs of firmware to be included in the set to be created
+  -h, --help                    help for firmware-set
+      --labels stringToString   Labels to assign to the firmware set - 'vendor=foo,model=bar' (default [])
+      --name string             A name for the firmware set
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl create](mctl_create.md)	 - Create resources
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
