+++ docs/mctl_create_firmware.md
@@ -1,0 +1,27 @@
+## mctl create firmware
+
+Create firmware
+
+```
+mctl create firmware [flags]
+```
+
+### Options
+
+```
+      --from-file string   YAML file with firmware configuration data
+  -h, --help               help for firmware
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl create](mctl_create.md)	 - Create resources
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
