+++ docs/mctl_delete.md
@@ -1,0 +1,28 @@
+## mctl delete
+
+Delete resources
+
+```
+mctl delete [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for delete
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl](mctl.md)	 - mctl is a CLI utility to interact with metal toolbox services
+* [mctl delete condition](mctl_delete_condition.md)	 - delete server condition
+* [mctl delete firmware-set](mctl_delete_firmware-set.md)	 - Delete a firmware set
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
