+++ docs/mctl_delete_condition.md
@@ -1,0 +1,28 @@
+## mctl delete condition
+
+delete server condition
+
+```
+mctl delete condition [flags]
+```
+
+### Options
+
+```
+  -h, --help               help for condition
+      --kind string        condition kind
+      --server-id string   server UUID
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl delete](mctl_delete.md)	 - Delete resources
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
