+++ docs/mctl_delete_firmware-set.md
@@ -1,0 +1,27 @@
+## mctl delete firmware-set
+
+Delete a firmware set
+
+```
+mctl delete firmware-set [flags]
+```
+
+### Options
+
+```
+  -h, --help          help for firmware-set
+      --uuid string   UUID of firmware set to be deleted
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl delete](mctl_delete.md)	 - Delete resources
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
