+++ docs/mctl_edit.md
@@ -1,0 +1,27 @@
+## mctl edit
+
+Edit resources
+
+```
+mctl edit [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for edit
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl](mctl.md)	 - mctl is a CLI utility to interact with metal toolbox services
+* [mctl edit firmware-set](mctl_edit_firmware-set.md)	 - Edit a firmware set
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
