+++ docs/mctl_gendocs.md
@@ -1,0 +1,26 @@
+## mctl gendocs
+
+Generate markdown docs for mctl CLI
+
+```
+mctl gendocs [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for gendocs
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl](mctl.md)	 - mctl is a CLI utility to interact with metal toolbox services
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
