+++ docs/mctl_generate.md
@@ -1,0 +1,26 @@
+## mctl generate
+
+Generate CLI docs
+
+```
+mctl generate [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for generate
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl](mctl.md)	 - mctl is a CLI utility to interact with metal toolbox services
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
