+++ docs/mctl_get.md
@@ -1,0 +1,33 @@
+## mctl get
+
+Get resource
+
+```
+mctl get [flags]
+```
+
+### Options
+
+```
+  -h, --help            help for get
+  -o, --output string   {json|text} (default "json")
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl](mctl.md)	 - mctl is a CLI utility to interact with metal toolbox services
+* [mctl get bios-config](mctl_get_bios-config.md)	 - Get bios configuration information for a server
+* [mctl get component](mctl_get_component.md)	 - get server components
+* [mctl get condition](mctl_get_condition.md)	 - get server condition
+* [mctl get firmware](mctl_get_firmware.md)	 - Get information for given firmware identifier
+* [mctl get firmware](mctl_get_firmware.md)	 - Get all firmware components on a server
+* [mctl get firmware-set](mctl_get_firmware-set.md)	 - Get information for given firmware set identifier
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
