+++ docs/mctl_get_component.md
@@ -1,0 +1,28 @@
+## mctl get component
+
+get server components
+
+```
+mctl get component [flags]
+```
+
+### Options
+
+```
+  -h, --help               help for component
+      --server-id string   server UUID
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+  -o, --output string   {json|text} (default "json")
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl get](mctl_get.md)	 - Get resource
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
