+++ docs/mctl_get_condition.md
@@ -1,0 +1,29 @@
+## mctl get condition
+
+get server condition
+
+```
+mctl get condition [flags]
+```
+
+### Options
+
+```
+  -h, --help               help for condition
+      --kind string        condition kind
+      --server-id string   server UUID
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+  -o, --output string   {json|text} (default "json")
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl get](mctl_get.md)	 - Get resource
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
