+++ docs/mctl_get_firmware-set.md
@@ -1,0 +1,29 @@
+## mctl get firmware-set
+
+Get information for given firmware set identifier
+
+```
+mctl get firmware-set [flags]
+```
+
+### Options
+
+```
+  -h, --help            help for firmware-set
+      --id string       firmware set UUID
+      --server string   server UUID
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+  -o, --output string   {json|text} (default "json")
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl get](mctl_get.md)	 - Get resource
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
