+++ docs/mctl_install.md
@@ -1,0 +1,27 @@
+## mctl install
+
+Install actions
+
+```
+mctl install [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for install
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl](mctl.md)	 - mctl is a CLI utility to interact with metal toolbox services
+* [mctl install firmware-set](mctl_install_firmware-set.md)	 - Install firmware set
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
