+++ docs/mctl_install_firmware-set.md
@@ -1,0 +1,30 @@
+## mctl install firmware-set
+
+Install firmware set
+
+```
+mctl install firmware-set [flags]
+```
+
+### Options
+
+```
+      --force            force install (skips firmware version check)
+  -h, --help             help for firmware-set
+      --id string        firmware set UUID
+      --server string    server UUID
+      --skip-bmc-reset   skip BMC reset before firmware install
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl install](mctl_install.md)	 - Install actions
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
