+++ docs/mctl_list.md
@@ -1,0 +1,30 @@
+## mctl list
+
+List resources
+
+```
+mctl list [flags]
+```
+
+### Options
+
+```
+  -h, --help          help for list
+      --output-json   Output listing as JSON
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl](mctl.md)	 - mctl is a CLI utility to interact with metal toolbox services
+* [mctl list condition](mctl_list_condition.md)	 - list server conditions by state
+* [mctl list firmware](mctl_list_firmware.md)	 - List firmware
+* [mctl list firmware-set](mctl_list_firmware-set.md)	 - List firmware
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
