+++ docs/mctl_list_condition.md
@@ -1,0 +1,29 @@
+## mctl list condition
+
+list server conditions by state
+
+```
+mctl list condition [flags]
+```
+
+### Options
+
+```
+  -h, --help               help for condition
+      --server-id string   server UUID
+      --state string       condition state
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --output-json     Output listing as JSON
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl list](mctl_list.md)	 - List resources
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
