+++ docs/mctl_list_firmware-set.md
@@ -1,0 +1,29 @@
+## mctl list firmware-set
+
+List firmware
+
+```
+mctl list firmware-set [flags]
+```
+
+### Options
+
+```
+  -h, --help            help for firmware-set
+      --model string    filter by server model
+      --vendor string   filter by server vendor
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --output-json     Output listing as JSON
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl list](mctl_list.md)	 - List resources
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
