+++ docs/mctl_list_firmware.md
@@ -1,0 +1,27 @@
+## mctl list firmware
+
+List firmware
+
+```
+mctl list firmware [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for firmware
+```
+
+### Options inherited from parent commands
+
+```
+      --config string   config file (default is $HOME/.mctl.yml)
+      --output-json     Output listing as JSON
+      --reauth          re-authenticate with oauth services
+```
+
+### SEE ALSO
+
+* [mctl list](mctl_list.md)	 - List resources
+
+###### Auto generated by spf13/cobra on 13-Jul-2023
